Changed Integer to Double as getLastDamage() now returns a Double.

+++ src/ro/thehunters/digi/recipeManager/flags/FlagExplode.java
@@ -17,21 +17,21 @@ import ro.thehunters.digi.recipeManager.RecipeManager;
 public class FlagExplode extends Flag
 {
     // Flag definition and documentation
-    
+
     private static final FlagType TYPE;
     protected static final String[] A;
     protected static final String[] D;
     protected static final String[] E;
-    
+
     static
     {
         TYPE = FlagType.EXPLODE;
-        
+
         A = new String[]
         {
             "{flag} <arguments or false>",
         };
-        
+
         D = new String[]
         {
             "Makes the workbench/furnace/player explode when recipe is crafted.",
@@ -45,7 +45,7 @@ public class FlagExplode extends Flag
             "  fail                   = (defualt not set) Explode if recipe failed as opposed to succeed.",
             "All arguments are optional and you can specify these arguments in any order.",
         };
-        
+
         E = new String[]
         {
             "{flag} // will explode when recipe succeeeds with power 2, 100% chance and breaks blocks",
@@ -53,19 +53,19 @@ public class FlagExplode extends Flag
             "{flag} fail | power 2 | chance 75% // will explode 75% of the time when recipe fails",
         };
     }
-    
+
     // Flag code
-    
+
     private float power = 2.0f;
     private boolean fire = false;
     private boolean noBreak = false;
     private byte noDamage = 0;
     private boolean failure = false;
-    
+
     public FlagExplode()
     {
     }
-    
+
     public FlagExplode(FlagExplode flag)
     {
         power = flag.power;
@@ -74,79 +74,79 @@ public class FlagExplode extends Flag
         noDamage = flag.noDamage;
         failure = flag.failure;
     }
-    
+
     @Override
     public FlagExplode clone()
     {
         return new FlagExplode(this);
     }
-    
+
     @Override
     public FlagType getType()
     {
         return TYPE;
     }
-    
+
     public float getPower()
     {
         return power;
     }
-    
+
     public void setPower(float power)
     {
         this.power = power;
     }
-    
+
     public boolean getFire()
     {
         return fire;
     }
-    
+
     public void setFire(boolean fire)
     {
         this.fire = fire;
     }
-    
+
     public boolean getFailure()
     {
         return failure;
     }
-    
+
     public void setFailure(boolean failure)
     {
         this.failure = failure;
     }
-    
+
     public boolean getNoBreak()
     {
         return noBreak;
     }
-    
+
     public void setNoBreak(boolean noBreak)
     {
         this.noBreak = noBreak;
     }
-    
+
     public boolean isNoDamageEnabled()
     {
         return noDamage > 0;
     }
-    
+
     public boolean isNoDamageSelf()
     {
         return noDamage == 2;
     }
-    
+
     public void setNoDamage(boolean enable)
     {
         setNoDamage(enable, false);
     }
-    
+
     public void setNoDamage(boolean enable, boolean self)
     {
         noDamage = (byte)(enable ? (self ? 2 : 1) : 0);
     }
-    
+
     @Override
     protected boolean onParse(String value)
     {
@@ -154,13 +154,13 @@ public class FlagExplode extends Flag
         {
             return true; // accepts null value
         }
-        
+
         String[] args = value.toLowerCase().split("\\|");
-        
+
         for(String arg : args)
         {
             arg = arg.trim().toLowerCase();
-            
+
             if(arg.equals("fire"))
             {
                 setFire(true);
@@ -176,13 +176,13 @@ public class FlagExplode extends Flag
             else if(arg.startsWith("nodamage"))
             {
                 value = arg.substring("nodamage".length()).trim();
-                
+
                 setNoDamage(true, value.equals("self"));
             }
             else if(arg.startsWith("power"))
             {
                 value = arg.substring("power".length()).trim();
-                
+
                 try
                 {
                     setPower(Float.valueOf(value));
@@ -198,10 +198,10 @@ public class FlagExplode extends Flag
                 ErrorReporter.warning("Flag " + getType() + " has unknown argument: " + arg);
             }
         }
-        
+
         return true;
     }
-    
+
     @Override
     protected void onCrafted(final Args a)
     {
@@ -214,38 +214,38 @@ public class FlagExplode extends Flag
             }
         }.runTaskLater(RecipeManager.getPlugin(), 1);
     }
-    
+
     private void boom(Args a)
     {
-        
+
         if(!a.hasLocation())
         {
             a.addCustomReason("Need a location!");
             return;
         }
-        
+
         if(failure && !a.hasResult())
         {
             a.addCustomReason("Needs a result!");
             return;
         }
-        
+
         boolean failed = (failure ? a.result().getType() == Material.AIR : failure);
-        
+
         if(failure == failed)
         {
-            Map<LivingEntity, Integer> entities = new HashMap<LivingEntity, Integer>();
+            Map<LivingEntity, Double> entities = new HashMap<LivingEntity, Double>();
             Location loc = a.location();
             World world = loc.getWorld();
             double x = loc.getX() + 0.5;
             double y = loc.getY() + 0.5;
             double z = loc.getZ() + 0.5;
-            
+
             if(isNoDamageEnabled())
             {
                 double distanceSquared = power * 2.0;
                 distanceSquared *= distanceSquared;
-                
+
                 if(isNoDamageSelf())
                 {
                     if(!a.hasPlayer())
@@ -256,7 +256,7 @@ public class FlagExplode extends Flag
                     {
                         Player p = a.player();
                         Location l = p.getLocation();
-                        
+
                         if(l.distanceSquared(loc) <= distanceSquared)
                         {
                             entities.put(p, p.getLastDamage());
@@ -278,10 +278,10 @@ public class FlagExplode extends Flag
                     }
                 }
             }
-            
+
             world.createExplosion(x, y, z, getPower(), getFire(), !getNoBreak());
-            
-            for(Entry<LivingEntity, Integer> e : entities.entrySet())
+
+            for(Entry<LivingEntity, Double> e : entities.entrySet())
             {
                 e.getKey().setNoDamageTicks(0);
                 e.getKey().setLastDamage(e.getValue());
